+++ .vscode/settings.json
@@ -1,3 +1,6 @@
 {
-    "python.pythonPath": "C:\\Users\\Alex\\.virtualenvs\\py-practice-lsRS_BrK\\scripts\\python.exe"
+    "python.pythonPath": "C:\\Users\\Alex\\.virtualenvs\\compiler-iEKU_aLA\\scripts\\python.exe"
+    ,"python.unitTest.unittestEnabled": true
+    ,"python.unitTest.unittestArgs": ["-s", "tests", "-p", "test_*.py"]
+
 }
