+++ Pipfile
@@ -11,3 +11,4 @@ name = "pypi"
 
 [dev-packages]
 
+rope = "*"
