--- program.txt
@@ -1,1 +1,0 @@
-53 + 2
