+++ tests/program.txt
@@ -1,0 +1,1 @@
+53 + 2
