+++ tests/test_compiler.py
@@ -1,0 +1,15 @@
+import unittest
+import context
+import compiler
+
+
+class TestLex(unittest.TestCase):
+
+    def test(self):
+        expected = ['53', '+', '2', '\n']
+        actual = compiler.lex('tests/program.txt')
+        self.assertEqual(expected, actual)
+
+
+class TestParser(unittest.TestCase):
+    pass
